refactor(router): pull-down 组件逻辑移动到 runtime
