+++ packages/taro-router/src/router.ts
@@ -2,12 +2,11 @@
 import UniversalRouter, { Routes } from 'universal-router'
 import { AppConfig, PageConfig } from '@tarojs/taro'
 import { LocationListener, LocationState } from 'history'
-import { createReactApp, createPageConfig, Current, createVueApp, PageInstance, eventCenter } from '@tarojs/runtime'
+import { createReactApp, createPageConfig, Current, createVueApp, PageInstance, eventCenter, CurrentReconciler } from '@tarojs/runtime'
 import { qs } from './qs'
 import { history } from './history'
 import { stacks } from './stack'
 import { init } from './init'
-import { createPullDownRefresh, R, V } from './pull-down'
 
 export interface Route extends PageConfig {
   path: string
@@ -98,7 +97,7 @@ export function createRouter (
   App,
   config: RouterConfig,
   type: 'react' | 'vue' | 'nerv',
-  framework: R | V,
+  framework,
   reactdom
 ) {
   init(config)
@@ -116,7 +115,7 @@ export function createRouter (
   }
 
   const router = new UniversalRouter(routes)
-  const app = type === 'vue' ? createVueApp(App, framework as V, config) : createReactApp(App, framework as R, reactdom, config)
+  const app = type === 'vue' ? createVueApp(App, framework, config) : createReactApp(App, framework, reactdom, config)
   app.onLaunch!()
 
   const render: LocationListener<LocationState> = async (location, action) => {
@@ -159,7 +158,7 @@ export function createRouter (
     if (shouldLoad) {
       const el = element.default ?? element
       const page = createPageConfig(
-        enablePullDownRefresh ? createPullDownRefresh(el, type, location.pathname, framework) : el,
+        enablePullDownRefresh ? CurrentReconciler.createPullDownComponent?.(el, location.pathname, framework) : el,
         location.pathname
       )
       loadPage(page)
