+++ packages/taro-runtime/src/index.ts
@@ -19,3 +19,4 @@ export { nextTick } from './next-tick'
 export { hydrate } from './hydrate'
 export * from './emitter/emitter'
 export { raf as requestAnimationFrame, caf as cancelAnimationFrame, now } from './bom/raf'
+export { Reconciler, CurrentReconciler } from './reconciler'
